# production-service/production-swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/machine/{id}": {
      "get": {
        "operationId": "MachineController_findOneMachine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Machine"
        ]
      },
      "delete": {
        "operationId": "MachineController_deleteOneMachine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Machine"
        ]
      }
    },
...
    "/supply-chain": {
      "get": {
        "operationId": "SupplyChainController_findAllSupplyChains",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Supply Chain"
        ]
      },
      "post": {
        "operationId": "SupplyChainController_createSupplyChain",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplyChainDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Supply Chain"
        ]
      }
    },
    "/supply-chain/{id}": {
      "get": {
        "operationId": "SupplyChainController_findSupplyChain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Supply Chain"
        ]
      },
      "patch": {
        "operationId": "SupplyChainController_updateSupplyChain",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplyChainDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Supply Chain"
        ]
      }
    },
    "/supply-chain/many-supply-chain": {
      "post": {
        "operationId": "SupplyChainController_createManySupplyChain",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManySupplyChainDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Supply Chain"
        ]
      },
      "patch": {
        "operationId": "SupplyChainController_updateManySupplyChain",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManySupplyChainDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Supply Chain"
        ]
      }
    }
  },
  "info": {
    "title": "Production API Service - Namkin X SUPINFO",
    "description": "<a href=\"/api/v1/production/production-swagger-spec.json\" target=\"_blank\">http://localhost:3000/api/v1/production/production-swagger-spec.json</a>",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateSupplyChainDto": {
        "type": "object",
        "properties": {
          "timeOfProduction": {
            "type": "string",
            "description": "Date of production for a given piece in timestamp (epoch)",
            "example": "10/10/2023"
          },
          "order": {
            "type": "number",
            "description": "Id of a piece order.",
            "example": 5
          },
          "var1": {
            "type": "number"
          },
          "var2": {
            "type": "string"
          },
          "var3": {
            "type": "string"
          },
          "var4": {
            "type": "number"
          },
          "var5": {
            "type": "boolean",
            "description": "State of a damaged piece.",
            "example": false
          },
          "machineIds": {
            "description": "Array of machine Ids.",
            "example": [
              1,
              2,
              3,
              4
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "partIds": {
            "description": "Array of part information Ids.",
            "example": [
              10,
              20,
              30,
              40
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "timeOfProduction",
          "order",
          "var1",
          "var2",
          "var3",
          "var4",
          "var5",
          "machineIds",
          "partIds"
        ]
      },
      "CreateManySupplyChainDto": {
        "type": "object",
        "properties": {
          "supplyChains": {
            "description": "Collection of supply chain objects.",
            "example": [
              {
                "timeOfProduction": "10/12/2023",
                "order": 5,
                "var5": false,
                "machineIds": [
                  1,
                  4,
                  7
                ],
                "partIds": [
                  40,
                  45,
                  50
                ]
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSupplyChainDto"
            }
          }
        },
        "required": [
          "supplyChains"
        ]
      },
      "UpdateSupplyChainDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "UUID of a given supply chain.",
            "example": "cde7964c-819b-11ee-b962-0242ac120002"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateManySupplyChainDto": {
        "type": "object",
        "properties": {
          "supplyChains": {
            "description": "Array of supply chain update objects.",
            "example": [
              "cde7964c-819b-11ee-b962-0242ac120002"
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateSupplyChainDto"
            }
          }
        },
        "required": [
          "supplyChains"
        ]
      }
    }
  }
}